pyproject.toml
[project]
name = "kalshi-weather-bot"
version = "0.1.0"
description = "Weather prediction market bot for Kalshi KXHIGH temperature markets"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "sqlalchemy>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10",
    "scikit-learn>=1.4",
    "numpy>=1.26",
    "pandas>=2.2",
    "cryptography>=42.0",
    "rich>=13.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.3",
    "mypy>=1.8",
]

[project.scripts]
weather-bot = "scripts.run_bot:main"
backfill = "scripts.backfill_history:main"
backtest = "scripts.run_backtest:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
testpaths = ["tests"]
